+++ example/calculatePrice/src/main/java/io/github/hanbernate/jsonbom/example/DiscountOrchestrator.java
@@ -1,14 +1,10 @@
 package io.github.hanbernate.jsonbom.example;
 
 import io.github.hanbernate.jsonbom.api.Bom;
-import io.github.hanbernate.jsonbom.api.BomMapping;
 import io.github.hanbernate.jsonbom.api.BomOrValue;
-import io.github.hanbernate.jsonbom.api.ValueHandler;
 import io.github.hanbernate.jsonbom.example.repository.PromotionRepository;
 import io.github.hanbernate.jsonbom.example.repository.PromotionRepository.Promotion;
 import java.math.BigDecimal;
-import java.util.List;
-import lombok.Data;
 import reactor.core.publisher.Mono;
 
 public class DiscountOrchestrator {