+++ src/main/java/io/github/hanbernate/jsonbom/spring/SpringBeanUtil.java
@@ -13,7 +13,6 @@ import java.util.AbstractMap;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
 public class SpringBeanUtil implements BeanUtil {